Refractor Make packages Shortening
