+++ src/app/src/main/java/com/softwareengineering/restaurant/AdminPackage/AdminGridAdapter.java
@@ -1,4 +1,4 @@
-package com.softwareengineering.restaurant;
+package com.softwareengineering.restaurant.AdminPackage;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -7,6 +7,8 @@ import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import com.softwareengineering.restaurant.R;
 
 public class AdminGridAdapter extends BaseAdapter {
