+++ src/app/src/main/java/com/softwareengineering/restaurant/AdminPackage/AdminMainActivity.java
@@ -1,4 +1,4 @@
-package com.softwareengineering.restaurant;
+package com.softwareengineering.restaurant.AdminPackage;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,8 +6,9 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.AdapterView;
-import android.widget.Toast;
 
+import com.softwareengineering.restaurant.R;
+import com.softwareengineering.restaurant.StaffPackage.StaffsActivity;
 import com.softwareengineering.restaurant.databinding.ActivityAdminMainBinding;
 
 public class AdminMainActivity extends AppCompatActivity {
