+++ src/app/src/main/java/com/softwareengineering/restaurant/AdminPackage/EditStaffsActivity.java
@@ -1,4 +1,4 @@
-package com.softwareengineering.restaurant;
+package com.softwareengineering.restaurant.AdminPackage;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.content.ContextCompat;
@@ -14,6 +14,8 @@ import android.widget.RadioGroup;
 import android.widget.TextView;
 
 import com.google.firebase.auth.FirebaseAuth;
+import com.softwareengineering.restaurant.R;
+import com.softwareengineering.restaurant.Staffs;
 
 import java.util.Objects;
