+++ src/app/src/main/java/com/softwareengineering/restaurant/AdminPackage/ReportsActivity.java
@@ -1,4 +1,4 @@
-package com.softwareengineering.restaurant;
+package com.softwareengineering.restaurant.AdminPackage;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.content.ContextCompat;
@@ -9,35 +9,26 @@ import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
-import android.widget.AdapterView;
 import android.widget.ImageView;
-import android.widget.LinearLayout;
-import android.widget.ListAdapter;
-import android.widget.ListView;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 
-import com.softwareengineering.restaurant.databinding.ActivityStaffsBinding;
+import com.softwareengineering.restaurant.R;
+import com.softwareengineering.restaurant.StaffPackage.StaffsActivity;
 
-import java.util.ArrayList;
-
-public class StaffsActivity extends AppCompatActivity {
+public class ReportsActivity extends AppCompatActivity {
 
     private DrawerLayout drawerLayout;
     private ImageView topMenuImg;
     private TextView topMenuName;
     private RelativeLayout staffs, customers, menu, tables, reports, sales, account;
-    private ActivityStaffsBinding binding;
-    private LinearLayout addStaffs;
 
     @SuppressLint("SetTextI18n")
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        binding = ActivityStaffsBinding.inflate(getLayoutInflater());
-        setContentView(binding.getRoot());
+        setContentView(R.layout.activity_reports);
 
         drawerLayout = findViewById(R.id.adminDrawerLayout);
         topMenuImg = findViewById(R.id.topMenuImg);
@@ -49,58 +40,8 @@ public class StaffsActivity extends AppCompatActivity {
         reports = findViewById(R.id.reportsDrawer);
         sales = findViewById(R.id.salesDrawer);
         account = findViewById(R.id.accountDrawer);
-        addStaffs = findViewById(R.id.adminStaffsAdd);
 
-        setItemBackgroundColors(staffs);
-
-        // Set data for Staffs list
-        // TODO: Need to get all staff with roles in firestore database
-        String[] staffsName = {
-                "Alpha", "Beta", "Charlie", "Delta"
-        };
-
-        String[] staffsRole = {
-                "Waiter", "Cook", "Cashier", "Janitor"
-        };
-
-        String[] staffsEmail = {
-                "alpha@12345.com", "beta@12345.com", "charlie@12345.com", "delta@12345.com"
-        };
-
-        String[] staffsGender = {
-                "Male", "Female"
-        };
-
-        String[] staffsPhone = {
-                "0123456789"
-        };
-
-        String[] staffsUsername = {
-                "Default"
-        };
-
-        // Initialize Staffs list
-        ArrayList<Staffs> staffsArrayList = new ArrayList<>();
-
-        for (int i = 0; i < staffsName.length; i++) {
-
-            Staffs tempStaff = new Staffs(staffsName[i], staffsEmail[i], staffsPhone[0], staffsGender[i % 2], staffsRole[i], staffsUsername[0]);
-            staffsArrayList.add(tempStaff);
-
-        }
-
-        StaffsAdapter staffsAdapter = new StaffsAdapter(StaffsActivity.this, staffsArrayList);
-
-        binding.staffsListView.setAdapter(staffsAdapter);
-        binding.staffsListView.setClickable(true);
-        binding.staffsListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                Intent intent = new Intent(StaffsActivity.this, StaffsDetails.class);
-                intent.putExtra("staffs", staffsArrayList.get(position));
-                startActivity(intent);
-            }
-        });
+        setItemBackgroundColors(reports);
 
         topMenuImg.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -109,13 +50,13 @@ public class StaffsActivity extends AppCompatActivity {
             }
         });
 
-        topMenuName.setText("Staffs List");
+        topMenuName.setText("Reports");
 
         staffs.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 setItemBackgroundColors(staffs);
-                recreate();
+                redirectActivity(ReportsActivity.this, StaffsActivity.class);
             }
         });
 
@@ -123,7 +64,7 @@ public class StaffsActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 setItemBackgroundColors(customers);
-                redirectActivity(StaffsActivity.this, CustomersActivity.class);
+                redirectActivity(ReportsActivity.this, CustomersActivity.class);
             }
         });
 
@@ -131,7 +72,7 @@ public class StaffsActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 setItemBackgroundColors(menu);
-                redirectActivity(StaffsActivity.this, MenuActivity.class);
+                redirectActivity(ReportsActivity.this, MenuActivity.class);
             }
         });
 
@@ -139,7 +80,7 @@ public class StaffsActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 setItemBackgroundColors(tables);
-                redirectActivity(StaffsActivity.this, TablesActivity.class);
+                redirectActivity(ReportsActivity.this, TablesActivity.class);
             }
         });
 
@@ -147,7 +88,7 @@ public class StaffsActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 setItemBackgroundColors(reports);
-                redirectActivity(StaffsActivity.this, ReportsActivity.class);
+                recreate();
             }
         });
 
@@ -155,7 +96,7 @@ public class StaffsActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 setItemBackgroundColors(sales);
-                redirectActivity(StaffsActivity.this, SalesActivity.class);
+                redirectActivity(ReportsActivity.this, SalesActivity.class);
             }
         });
 
@@ -163,29 +104,10 @@ public class StaffsActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 setItemBackgroundColors(account);
-                redirectActivity(StaffsActivity.this, AccountActivity.class);
+                redirectActivity(ReportsActivity.this, AccountActivity.class);
             }
         });
 
-        // Handle Add Staffs
-        addStaffs.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(StaffsActivity.this, AddStaffsActivity.class);
-                startActivity(intent);
-            }
-        });
-
-        // Handle new created staff account
-        Staffs newStaffs = getIntent().getParcelableExtra("newStaffs");
-        if (newStaffs != null) {
-            staffsArrayList.add(newStaffs);
-            staffsAdapter.notifyDataSetChanged();
-            binding.staffsListView.setAdapter(staffsAdapter);
-            binding.staffsListView.setClickable(true);
-        }
-
-        staffsAdapter.notifyDataSetChanged();
     }
 
     private void setItemBackgroundColors(RelativeLayout selectedItem) {
