+++ src/app/src/main/res/layout/activity_add_staffs.xml
@@ -5,7 +5,7 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddStaffsActivity">
+    tools:context=".AdminPackage.AddStaffsActivity">
 
     <ScrollView
         android:layout_width="match_parent"
