+++ src/app/src/main/res/layout/activity_admin_main.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true"
-    tools:context=".AdminMainActivity">
+    tools:context=".AdminPackage.AdminMainActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
