+++ src/app/src/main/res/layout/activity_edit_staffs.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".EditStaffsActivity">
+    tools:context=".AdminPackage.EditStaffsActivity">
 
     <ScrollView
         android:layout_width="match_parent"
