+++ src/app/src/main/res/layout/activity_menu.xml
@@ -4,7 +4,7 @@
     android:id="@+id/adminDrawerLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MenuActivity">
+    tools:context=".AdminPackage.MenuActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
