+++ src/app/src/main/res/layout/activity_reports.xml
@@ -4,7 +4,7 @@
     android:id="@+id/adminDrawerLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ReportsActivity">
+    tools:context=".AdminPackage.ReportsActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
