+++ src/app/src/main/res/layout/activity_staffs.xml
@@ -4,7 +4,7 @@
     android:id="@+id/adminDrawerLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".StaffsActivity">
+    tools:context=".StaffPackage.StaffsActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
