+++ src/app/src/main/res/layout/activity_staffs_account.xml
@@ -7,7 +7,7 @@
     android:fitsSystemWindows="true"
     tools:openDrawer="start"
     android:id="@+id/staffsDrawerLayout"
-    tools:context=".StaffsAccountActivity">
+    tools:context=".StaffPackage.StaffsAccountActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
