+++ src/app/src/main/res/layout/activity_staffs_customers.xml
@@ -7,7 +7,7 @@
     android:fitsSystemWindows="true"
     tools:openDrawer="start"
     android:id="@+id/staffsDrawerLayout"
-    tools:context=".StaffsCustomersActivity">
+    tools:context=".StaffPackage.StaffsCustomersActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
