+++ src/app/src/main/res/layout/activity_staffs_details.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".StaffsDetails">
+    tools:context=".AdminPackage.StaffsDetails">
 
     <ScrollView
         android:layout_width="match_parent"
