+++ src/app/src/main/res/layout/activity_staffs_payment.xml
@@ -7,7 +7,7 @@
     android:fitsSystemWindows="true"
     tools:openDrawer="start"
     android:id="@+id/staffsDrawerLayout"
-    tools:context=".StaffsPaymentActivity">
+    tools:context=".StaffPackage.StaffsPaymentActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
