+++ src/app/src/main/res/layout/activity_staffs_reports.xml
@@ -7,7 +7,7 @@
     android:fitsSystemWindows="true"
     tools:openDrawer="start"
     android:id="@+id/staffsDrawerLayout"
-    tools:context=".StaffsReportsActivity">
+    tools:context=".StaffPackage.StaffsReportsActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
