+++ src/app/src/main/res/layout/activity_staffs_tables.xml
@@ -7,7 +7,7 @@
     android:fitsSystemWindows="true"
     tools:openDrawer="start"
     android:id="@+id/staffsDrawerLayout"
-    tools:context=".StaffsTablesActivity">
+    tools:context=".StaffPackage.StaffsTablesActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
