+++ src/app/src/main/res/layout/activity_tables.xml
@@ -5,7 +5,7 @@
     android:id="@+id/adminDrawerLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TablesActivity">
+    tools:context=".AdminPackage.TablesActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
